Split views.py into various modules
