+++ ap_src/ap_app/absences.py
@@ -1,0 +1,307 @@
+#get_absence_data
+#text_rules
+#profile_page
+#absence_delete
+#absence_recurring_delete
+#edit_recurring_absences
+#add_recurring
+
+import calendar
+import datetime
+import json
+import holidays
+import pycountry
+import pandas as pd
+import requests
+from datetime import timedelta
+
+from django.contrib import messages
+from django.contrib.auth import get_user_model
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from django.db.models.functions import Lower
+from django.http import HttpResponse, JsonResponse
+from django.shortcuts import redirect, render
+from django.urls import reverse, reverse_lazy
+from django.views.generic import CreateView, UpdateView
+
+from .teams import *
+from .objects import *
+from .teams import *
+
+from .forms import *
+from .models import (Absence, RecurringAbsences, Relationship, Role, Team,
+                     UserProfile)
+
+def get_absence_data(users, user_type):
+    data = {}
+    absence_content = []
+    total_absence_dates = {}
+    total_half_dates = {}
+    total_recurring_dates = {}
+    all_absences = {}
+    delta = datetime.timedelta(days=1)
+
+    for user in users:
+        # all the absences for the user
+        if user_type == 1:
+            user_username = user.user.username
+        else:
+            user_username = user.username
+
+        absence_info = Absence.objects.filter(Target_User_ID__username=user_username)
+        total_absence_dates[user_username] = []
+        total_recurring_dates[user_username] = []
+        total_half_dates[user_username] = []
+        all_absences[user_username] = []
+
+        # if they have any absences
+        if absence_info:
+            # mapping the absence content to keys in dictionary
+            for x in absence_info:
+                absence_id = x.ID
+                absence_date_start = x.absence_date_start
+                absence_date_end = x.absence_date_end
+                dates = absence_date_start
+                if x.half_day == "NORMAL":
+                    while dates <= absence_date_end:
+                        total_absence_dates[user_username].append(dates)
+                        dates += delta
+                else:
+                    total_half_dates[user_username].append(
+                        {
+                            "date": absence_date_start,
+                            "type": x.half_day
+                        }
+                    )
+
+                absence_content.append(
+                    {
+                        "ID": absence_id,
+                        "absence_date_start": absence_date_start,
+                        "absence_date_end": absence_date_end,
+                        "dates": total_absence_dates[user_username],
+                    }
+                )
+
+            # for each user it maps the set of dates to a dictionary key labelled as the users name
+            all_absences[user_username] = absence_content
+
+        recurring = RecurringAbsences.objects.filter(Target_User_ID__username=user_username)
+
+        if recurring:
+            for recurrence_ in recurring:
+                dates = recurrence_.Recurrences.occurrences(
+                    dtend=datetime.datetime.strptime(
+                        str(datetime.datetime.now().year + 2), "%Y"
+                    )
+                )
+
+                for x in list(dates)[:-1]:
+                    time_const = "23:00:00"
+                    time_var = datetime.datetime.strftime(x, "%H:%M:%S")
+                    if time_const == time_var:
+                        x += timedelta(days=1)
+                    
+                    #print(RecurringException.objects.filter(Target_User_ID__username=user_username, Exception_Start=x).count())
+                    if RecurringException.objects.filter(Target_User_ID__username=user_username, Exception_Start=x).count() == 0:
+                        total_recurring_dates[user_username].append(x)
+                # TODO: add auto deleting for recurring absences once last date of absences in before now
+                # if x < datetime.datetime.now():
+                #    pass
+
+    data["recurring_absence_dates"] = total_recurring_dates
+    data["all_absences"] = all_absences
+    data["absence_dates"] = total_absence_dates
+    data["half_days_data"] = total_half_dates
+    data["users"] = users
+    return data
+
+def text_rules(user):
+    recurring_absences = RecurringAbsences.objects.filter(Target_User_ID=user).values(
+        "Recurrences", "ID"
+    )
+    rec_absences = {}
+
+    for x in recurring_absences:
+        absence_ = x["Recurrences"]
+        if absence_:
+            rec_absences[x["ID"]] = []
+        if absence_.exdates:
+            for y in absence_.exdates:
+                rec_absences[x["ID"]].append(
+                    "Excluding Date: " + (y + timedelta(days=1)).strftime("%A,%d %B,%Y")
+                )
+        if absence_.rdates:
+            for y in absence_.rdates:
+                rec_absences[x["ID"]].append(
+                    "Date: " + (y + timedelta(days=1)).strftime("%A,%d %B,%Y")
+                )
+        if absence_.rrules:
+            for y in absence_.rrules:
+                rec_absences[x["ID"]].append("Rule: " + str(y.to_text()))
+
+        if absence_.exrules:
+            for y in absence_.exrules:
+                rec_absences[x["ID"]].append("Excluding Rule: " + str(y.to_text()))
+
+    return rec_absences
+
+
+# Profile page
+@login_required
+def profile_page(request):
+    if request.method == "POST":
+        form = SwitchUser(
+            request.POST,
+            initial={"user": request.user},
+        )
+        form.fields["user"].queryset = UserProfile.objects.filter(
+            edit_whitelist=request.user
+        )
+        users = UserProfile.objects.filter(edit_whitelist=request.user)
+        rec_absences = text_rules(request.user)
+
+        if form.is_valid():
+            absence_user = form.cleaned_data["user"].user
+
+            absences = Absence.objects.filter(Target_User_ID=absence_user)
+            rec_absences = text_rules(absence_user)
+            return render(
+                request,
+                "ap_app/profile.html",
+                {
+                    "form": form,
+                    "message": "Successfully switched user",
+                    "absences": absences,
+                    "users": users,
+                    "recurring_absences": rec_absences,
+                },
+            )
+    else:
+        absences = Absence.objects.filter(Target_User_ID=request.user.id)
+        rec_absences = text_rules(request.user)
+
+        users = UserProfile.objects.filter(edit_whitelist=request.user)
+        form = SwitchUser()
+        form.fields["user"].queryset = users
+        form.fields["user"].initial = request.user
+
+        return render(
+            request,
+            "ap_app/profile.html",
+            {
+                "absences": absences,
+                "users": users,
+                "form": form,
+                "recurring_absences": rec_absences,
+            },
+        )
+    
+@login_required
+def absence_delete(request, absence_id: int, user_id: int, team_id: int = 1):
+    try:
+        absence = Absence.objects.get(pk=absence_id)
+        absence.delete()
+    except Absence.DoesNotExist: 
+        pass
+    if request.user == User.objects.get(id = user_id):
+        return redirect("profile")
+    return redirect("edit_team_member_absence", team_id, user_id)
+
+
+@login_required
+def absence_recurring_delete(
+    request, absence_id: int, user_id: int, team_id: int = None
+):
+    absence = RecurringAbsences.objects.get(pk=absence_id)
+    user = request.user
+    absence.delete()
+    if user == absence.Target_User_ID:
+        return redirect("profile")
+    return redirect("edit_team_member_absence", team_id, user_id)
+
+
+class EditAbsence(UpdateView):
+    template_name = "ap_app/edit_absence.html"
+    model = Absence
+
+    # specify the fields
+    fields = ["absence_date_start", "absence_date_end"]
+
+    def get_success_url(self) -> str:
+        return reverse("profile")
+
+
+@login_required
+def edit_recurring_absences(request, pk):
+    absence = RecurringAbsences.objects.get(ID=pk)
+
+    if request.method == "POST":
+        form = RecurringAbsencesForm(request.POST, instance=absence)
+        form2 = TargetUserForm(request.POST, target_user=request.user)
+        form2.fields["target_user"].queryset = UserProfile.objects.filter(
+            edit_whitelist__in=[request.user]
+        )
+        rule = str(form["Recurrences"].value())
+
+        if not ("DAILY" in rule or "BY" in rule):
+            content = {
+                "form": form,
+                "form2": form2,
+                "message": "Must select a day/month",
+            }
+            return render(request, "ap_app/edit_recurring_absence.html", content)
+
+        if form2.is_valid():
+            absence.Target_User_ID = form2.cleaned_data["target_user"].user
+            absence.recurrences = form["Recurrences"].value()
+            absence.save()
+            return redirect("index")
+    else:
+        form = RecurringAbsencesForm(instance=absence)
+        
+        form2 = TargetUserForm(target_user=absence.Target_User_ID)
+        form2.fields["target_user"].queryset = UserProfile.objects.filter(
+            edit_whitelist__in=[absence.User_ID]
+        )
+
+    return render(
+        request, "ap_app/edit_recurring_absence.html", {"form": form, "form2": form2}
+    )
+
+@login_required
+def add_recurring(request) -> render:
+    if request.method == "POST":
+        form = RecurringAbsencesForm(request.POST)
+        form2 = TargetUserForm(request.POST, target_user=request.user)
+        form2.fields["target_user"].queryset = UserProfile.objects.filter(
+            edit_whitelist__in=[request.user]
+        )
+        rule = str(form["Recurrences"].value())
+
+
+        if not ("DAILY" in rule or "BY" in rule):
+            content = {
+                "form": form,
+                "form2": form2,
+                "message": "Must select a day/month",
+            }
+            return render(request, "ap_app/add_recurring_absence.html", content)
+
+        if form2.is_valid():
+            RecurringAbsences.objects.create(
+                Recurrences=form["Recurrences"].value(),
+                Target_User_ID=form2.cleaned_data["target_user"].user,
+                User_ID=request.user,
+            )
+            return redirect("index")
+    else:
+        form = RecurringAbsencesForm()
+        form2 = TargetUserForm(target_user=request.user)
+        form2.fields["target_user"].queryset = UserProfile.objects.filter(
+            edit_whitelist__in=[request.user]
+        )
+
+    content = {"form": form, "form2": form2}
+    return render(request, "ap_app/add_recurring_absence.html", content)
