+++ ap_src/ap_app/errors.py
@@ -1,0 +1,22 @@
+def handler404(request, exception):
+    context = {}
+    response = render(request, "404.html", context=context)
+    response.status_code = 404
+    return response
+
+def handler500(request):
+    context = {}
+    response = render(request, "500.html", context=context)
+    response.status_code = 500
+    return response
+
+def handler400(request, exception):
+    context = {}
+    response = render(request, "400.html", context=context)
+    response.status_code = 400
+    return response
+
+from django.shortcuts import render
+
+def my_view(request):
+    return render(request, "base.html")
