+++ ap_src/ap_app/views.py
@@ -1,6 +1,14 @@
-# TODO: Review functions and variables names and ensure that they
-#   ..  have meaningful names that represent what they do.
-
+#index
+#privacy_page
+#add
+#details_page
+#deleteuser
+#profile_settings
+#add_user
+#get_region_data
+#set_region
+#click_add
+#click_remove
 
 import calendar
 import datetime
@@ -22,16 +30,18 @@ from django.http import HttpResponse, JsonResponse, HttpRequest
 from django.shortcuts import redirect, render
 from django.urls import reverse, reverse_lazy
 from django.views.generic import CreateView, UpdateView
-from django.contrib.auth.models import User
 
-from collections import namedtuple
+from .teams import *
+from .objects import *
+from .teams import *
+from .calendarview import *
 
 from .forms import *
 from .models import (Absence, RecurringAbsences, Relationship, Role, Team,
-                     UserProfile, Status, RecurringException)
+                     UserProfile)
 
-env = environ.Env()
-environ.Env.read_env()
+User = get_user_model()
+
 
 def index(request) -> render:
     """Branched view.                       \n
@@ -75,594 +85,66 @@ def privacy_page(request, to_accept=False) -> render:
         return render(request, "ap_app/privacy.html")
     return all_calendar(request)
 
-
 class SignUpView(CreateView):
     form_class = UserCreationForm
     success_url = reverse_lazy("login")
     template_name = "registration/signup.html"
 
-
 @login_required
-def teams_dashboard(request) -> render:
-    rels = Relationship.objects.order_by(Lower("team__name")).filter(
-        user=request.user, status=Status.objects.get(status="Active")
-    )
-    invite_rel_count = Relationship.objects.filter(
-        user=request.user, status=Status.objects.get(status="Invited")
-    ).count()
-    
-    try:
-        token = (str(request.user) + "AbsencePlanner").encode()
-        encryption = hashlib.sha256(token).hexdigest()
-        r = requests.get(env("TEAM_DATA_URL") + "api/teams/?username={}".format(request.user.username), headers={"TEAMS-TOKEN": encryption})
-    except:
-        return render(
-        request,
-        "teams/dashboard.html",
-        {"rels": rels, "invite_count": invite_rel_count, "external_teams": False, "teamspage_active": True})
-    if r.status_code == 200:
-        if len(r.json()) == 0 :
-            external_teams_data = False
-        else:
-            external_teams_data = r.json()
-    else:
-        external_teams_data = False
-    return render(
-        request,
-        "teams/dashboard.html",
-        {"rels": rels, "invite_count": invite_rel_count, "external_teams": external_teams_data, "teamspage_active": True},
-    )
-
-
-@login_required
-def create_team(request) -> render:
-    if request.method == "POST":
-        form = CreateTeamForm(request.POST)
-        if form.name_similarity():
-            # TODO: write code to tell the user that their team name is similar and to give them
-            # options to change the team name.
-            return HttpResponse(
-                "Debug: Did not create form because the name is too similar to another team name"
-            )
-
-        if form.is_valid():
-            form.save()
-            # Gets the created team and "Owner" Role and creates a Link between
-            # the user and their team
-            created_team = Team.objects.get(name=form.cleaned_data["name"])
-            assign_role = Role.objects.get(role="Owner")
-            Relationship.objects.create(
-                user=request.user,
-                team=created_team,
-                role=assign_role,
-                status=Status.objects.get(status="Active"),
-            )
-            return redirect("/teams/", {"message": "Team successfully created."})
-    else:
-        form = CreateTeamForm()
-    teams = Team.objects.all()
-    existing_teams = ""
-    existing_teams_ids = ""
-    for i, team in enumerate(teams):
-        existing_teams += team.name
-        existing_teams_ids += str(team.id)
-        if i != len(teams) - 1:
-            existing_teams += ","
-            existing_teams_ids += ","
-
-    return render(
-        request,
-        "teams/create_team.html",
-        {
-            "form": form,
-            "existing_teams": existing_teams,
-            "existing_teams_ids": existing_teams_ids,
-        },
-    )
-
-
-@login_required
-def join_team(request) -> render:
-    """Renders page with all teams the user is not currently in"""
-    user_teams = []
-    all_user_teams = Relationship.objects.filter(
-        user=request.user, status=Status.objects.get(status="Active")
-    )
-    for teams in all_user_teams:
-        user_teams.append(teams.team)
-    all_teams = Team.objects.all().exclude(relationship__user=request.user.id)
-    all_teams_filtered = []
-
-    # Filtering by team name
-    if "teamName" in request.GET:
-        for team in all_teams:
-            if request.GET["teamName"].lower() in team.name.lower():
-                all_teams_filtered.append(team)
-    else:
-        all_teams_filtered = all_teams
-
-    return render(
-        request,
-        "teams/join_team.html",
-        {"all_teams": all_teams_filtered, "joined_teams": user_teams},
-    )
-
-
-@login_required
-def joining_team_process(request, id, role):
-    find_team = Team.objects.get(id=id)
-    find_role = Role.objects.get(role=role)
-    rels = Relationship.objects.filter(
-        user=request.user,
-        role=Role.objects.get(role="Member"),
-        status=Status.objects.get(status="Active"),
-    )
-    rels2 = Relationship.objects.filter(
-        user=request.user,
-        role=Role.objects.get(role="Owner"),
-        status=Status.objects.get(status="Active"),
-    )
-    existing_rels = Relationship.objects.order_by(Lower("team__name")).filter(
-        user=request.user, status=Status.objects.get(status="Active")
-    )
-    invite_rel_count = Relationship.objects.filter(
-        user=request.user, status=Status.objects.get(status="Invited")
-    ).count()
-    
-    if (rels or rels2) and role == "Member":
-        return render(
-        request,
-        "teams/dashboard.html",
-        {"rels": existing_rels, "invite_count": invite_rel_count, "teamspage_active": True, "message" : "You are already part of one team", "message_type":"is-danger"},
-    )
-    new_rel = Relationship.objects.create(
-        user=request.user,
-        team=find_team,
-        role=find_role,
-        status=Status.objects.get(status="Pending"),
-    )
-    if not find_team.private:
-        Relationship.objects.filter(id=new_rel.id).update(
-            status=Status.objects.get(status="Active")
-        )
-        #leader = Relationship.objects.get(team=id, role=Role.objects.get(role="Owner"))
-        #userprofile = UserProfile.objects.get(user=request.user)
-        #userprofile.edit_whitelist.add(leader.user)
-        #userprofile.save()
-    return redirect("dashboard")
-
-
-def team_invite(request, team_id, user_id, role):
-    find_team = Team.objects.get(id=team_id)
-    find_user = User.objects.get(id=user_id)
-    find_role = Role.objects.get(role=role)
-
-    test = Relationship.objects.filter(team=find_team)
-
-    # Boolean determines if viewer is in this team trying to invite others
-    user_acceptable = False
-    for rel in test:
-        if rel.user == request.user and str(rel.role) == "Owner":
-            user_acceptable = True
-            break
-
-    if str(request.user.id) != str(user_id) and user_acceptable:
-        Relationship.objects.create(
-            user=find_user,
-            team=find_team,
-            role=find_role,
-            status=Status.objects.get(status="Invited"),
-        )
-        return redirect(f"/teams/calendar/{find_team.id}")
-    # Else user is manipulating the URL making non-allowed invites - (Therefore doesn't create a relationship)
-
-    return redirect("dashboard")
-
-
-@login_required
-def view_invites(request):
-    all_invites = Relationship.objects.filter(
-        user=request.user, status=Status.objects.get(status="Invited")
-    )
-    return render(request, "teams/invites.html", {"invites": all_invites})
-
-
-@login_required
-def leave_team(request, id):
-    find_relationship = Relationship.objects.get(id=id)
-    find_relationship.custom_delete()
-    team_cleaner(find_relationship)
-    return redirect("dashboard")
-
-
-def team_cleaner(rel):
-    """Detects if a team is empty and deletes it if it is."""
-    team = Team.objects.get(id=rel.team.id)
-    all_team_relationships = Relationship.objects.filter(team=team)
-    if all_team_relationships.count() == 0:
-        team.delete()
-
-
-@login_required
-def team_misc(request, id):
-    """Teams Miscellaneous/Notes page"""
-    if is_member(user=request.user, team_id=id):
-        team = Team.objects.get(id=id)
-
-        notes = CreateTeamForm(instance=team)
-
-        # TODO: Add a field to each team with a notes section - (for now it's just the teams description)
-
-        if "value" in request.GET:
-            team.notes = request.GET["value"]
-            team.save()
-
-        desc = team.description
-        notes = team.notes
-
-        return render(
-            request, "teams/misc.html", {"team": team, "desc": desc, "notes": notes}
-        )
-    return redirect("dashboard")
-
-
-@login_required
-def team_settings(request, id):
-    """Checks to see if user is the owner and renders the Setting page"""
-    team = Team.objects.get(id=id)
-    
-    if is_owner(user=request.user, team_id=id): 
-        all_pending_relations = Relationship.objects.filter(
-            team=id, status=Status.objects.get(status="Pending")
-        )
-        return render(
-            request,
-            "teams/settings.html",
-            {
-                "user": request.user,
-                "team": team,
-                "pending_rels": all_pending_relations,
-                "Team_users": team.users,
-                "member": Role.objects.get(role="Member"),
-                "coowner": Role.objects.get(role="Co-Owner"),
-                "follower": Role.objects.get(role="Follower"),
-                "owner": Role.objects.get(role="Owner"),
-            },
-        )
-    return redirect("dashboard")
-
-def edit_team_member_absence(request, id, user_id) -> render:
-    """Checks to see if user is the owner and renders the Edit absences page for that user"""
-    team = Team.objects.get(id=id)
-        
-    if is_owner(user=request.user, team_id=id):    
-        all_pending_relations = Relationship.objects.filter(
-            team=id, status=Status.objects.get(status="Pending")
-        )
-        target_user = User.objects.get(id=user_id)
-        absences = Absence.objects.filter(Target_User_ID=target_user.id)
-        rec_absences = text_rules(target_user)
-        return render(
-            request,
-            "teams/edit_absences.html",
-            {
-                "team": team,
-                "user": target_user,
-                "absences": absences,
-                "recurring_absences": rec_absences,
-            },
-        )
-    return redirect("dashboard")
-
-
-@login_required
-def promote_team_member(request, id, user_id):
-    """Checks to see if user is the owner and renders the Setting page"""
-    team = Team.objects.get(id=id)
-    user_relation = Relationship.objects.get(team=id, user=request.user)
-    if user_relation.role.role != "Owner":
-        return redirect("dashboard")
-    target_user = User.objects.get(id=user_id)
-    all_pending_relations = Relationship.objects.filter(
-        team=id, status=Status.objects.get(status="Pending")
-    )
-    current_relationship = Relationship.objects.get(team=team, user=target_user)
-    if current_relationship.role == Role.objects.get(role="Member"):
-        current_relationship.role = Role.objects.get(role="Co-Owner")
-        current_relationship.save()
-
-    return redirect(team_settings, team.id)
-
-
-@login_required
-def demote_team_member(request, id, user_id):
-    """Checks to see if user is the owner and renders the Setting page"""
-    team = Team.objects.get(id=id)
-    user_relation = Relationship.objects.get(team=id, user=request.user)
-    if user_relation.role.role != "Owner":
-        return redirect("dashboard")
-    target_user = User.objects.get(id=user_id)
-    all_pending_relations = Relationship.objects.filter(
-        team=id, status=Status.objects.get(status="Pending")
-    )
-    current_relationship = Relationship.objects.get(team=team, user=target_user)
-    if current_relationship.role == Role.objects.get(role="Co-Owner"):
-        current_relationship.role = Role.objects.get(role="Member")
-        current_relationship.save()
-
-    return redirect(team_settings, team.id)
-
-@login_required
-def remove_team_member(request, id, user_id):
-    """Removes a member from a team."""
-    team = Team.objects.get(id=id)
-    user_relation = Relationship.objects.get(team=id, user=request.user)
-    if user_relation.role.role != "Owner":
-        return redirect("dashboard") #Checks if the user is the owner and redirects to the dashboard if they aren't
-    target_user = User.objects.get(id=user_id) #Gets the user to be removed
-    target_relation = Relationship.objects.get(team=team, user=target_user) #Gets the target's relationship to the team
-    target_relation.custom_delete() #Deletes the relationship from the team, removing the user
-
-    return redirect(team_settings, team.id) #Redirects user back to the settings page
-
-
-def joining_team_request(request, id, response):
-
-    find_rel = Relationship.objects.get(id=id)
-    
-
-    if response == "accepted":
-        state_response = Status.objects.get(status="Active")
-        Relationship.objects.filter(id=find_rel.id).update(status=state_response)
-        return redirect("team_settings", find_rel.team.id)
-    elif response == "nonactive":
-        return redirect("leave_team", id)
-    return redirect("dashboard")
-
-#JC - Add an absence via the form
-@login_required
-def manual_add(request:HttpRequest) -> render:
-    #JC - POST request
+def example(request):
     if request.method == "POST":
         form = AbsenceForm(
-            request.POST, 
+            request.POST,
             user=request.user,
             initial={
                 "start_date": datetime.datetime.now(),
-                "end_date": lambda: datetime.datetime.now().date() + datetime.timedelta(days=1)
-            }
+                "end_date": lambda: datetime.datetime.now().date()
+                + datetime.timedelta(days=1),
+            },
         )
         form.fields["user"].queryset = UserProfile.objects.filter(
             edit_whitelist__in=[request.user]
         )
 
-        #JC - Create absence
         if form.is_valid():
-            absence = Absence()
-            absence.absence_date_start = form.cleaned_data["start_date"]
-            absence.absence_date_end = form.cleaned_data["end_date"]
-            absence.User_ID = request.user
-            absence.Target_User_ID = form.cleaned_data["user"].user
-
-            #JC - Check if the dates overlap with an existing absence. 
-            valid = True
-            Range = namedtuple('Range', ['start', 'end'])
-            r1 = Range(start=absence.absence_date_start, end=absence.absence_date_end)
-            for x in Absence.objects.filter(Target_User_ID=form.cleaned_data["user"].user.id):
-                r2 = Range(start=x.absence_date_start, end=x.absence_date_end)
-                delta = (min(r1.end, r2.end) -max(r1.start, r2.start)).days + 1
-                overlapp = max(0, delta)
-                if overlapp == 1:
-                    valid = False
-            
-            if valid:
-                absence.save()
-                return redirect("/")
-            else:
-                return render(request, "ap_app/add_absence.html", {
+            obj = Absence()
+            obj.absence_date_start = request.POST.get("start_date")
+            obj.absence_date_end = request.POST.get("end_date")
+            obj.absence_date_start = form.cleaned_data["start_date"]
+            obj.absence_date_end = form.cleaned_data["end_date"]
+            obj.request_accepted = False  # TODO
+            obj.User_ID = request.user
+            obj.Target_User_ID = form.cleaned_data["user"].user
+            message = "Absence successfully created"
+            msg_type = "is-success"
+            absence_valid = True
+            for x in Absence.objects.filter(User_ID=request.user.id):
+                if (obj.absence_date_start >= x.absence_date_start and obj.absence_date_end <= x.absence_date_end) \
+                    or x.absence_date_start == obj.absence_date_end or x.absence_date_end == x.absence_date_start:
+                        message = "Absence already created"
+                        msg_type = "is-danger"
+                        absence_valid = False
+            if absence_valid:
+                obj.save()
+            return render(
+                request,
+                "ap_app/add_absence.html",
+                {
                     "form": form,
-                    "message": "The absence conflicts with an existing absence",
-                    "message_type": "is-danger"
-                })
-
-    #JC - GET request
+                    "message": message,
+                    "message_type": msg_type,
+                    "add_absence_active": True,
+                },
+            )
+            # redirect to success page
     else:
         form = AbsenceForm(user=request.user)
-        #JC - Allow users to edit others absence if they have permission
         form.fields["user"].queryset = UserProfile.objects.filter(
             edit_whitelist__in=[request.user]
         )
-    
-    content = {"form": form}
+
+    content = {"form": form, "add_absence_active": True}
     return render(request, "ap_app/add_absence.html", content)
-
-#Add an absence when clicking on the calendar
-@login_required
-def click_add(request):
-    if request.method == "POST":
-        json_data=json.loads(request.body)
-        if UserProfile.objects.filter(user__username=json_data["username"]).exists():
-            perm_list = UserProfile.objects.filter(user__username=json_data["username"])[0].edit_whitelist.all()
-        else:
-            perm_list = [UserProfile.objects.get(user=request.user)]
-        if request.user in perm_list:
-            date = datetime.datetime.strptime(json_data["date"], "%Y-%m-%d").date()
-            #This will add a half
-            if json_data["half_day"] == True:
-                absence = Absence()
-                absence.absence_date_start = json_data['date']
-                absence.absence_date_end = json_data['date']
-                absence.Target_User_ID_id = User.objects.get(username=json_data["username"]).id
-                absence.User_ID = request.user
-                if json_data["half_day_time"] == "M":
-                    absence.half_day = "MORNING"
-                elif json_data["half_day_time"] == "A":
-                    absence.half_day = "AFTERNOON"
-                absence.save()
-                return JsonResponse({})
-            else:
-                def non_connected():
-                    absence = Absence()
-                    absence.absence_date_start = json_data['date']
-                    absence.absence_date_end = json_data['date']
-                    absence.Target_User_ID_id = User.objects.get(username=json_data["username"]).id
-                    absence.User_ID = request.user
-                    absence.save()
-                    return absence
-                date = datetime.datetime.strptime(json_data["date"], "%Y-%m-%d").date()
-                absence = None
-                if date - timedelta(days=1) in Absence.objects.filter(Target_User_ID__username=json_data["username"]).values_list("absence_date_end", flat=True) \
-                    and date + timedelta(days=1) in Absence.objects.filter(Target_User_ID__username=json_data["username"]).values_list("absence_date_start", flat=True):
-                    ab_1 = Absence.objects.filter(Target_User_ID__username=json_data["username"], absence_date_start=date+timedelta(days=1))[0]
-                    ab_2 = Absence.objects.filter(Target_User_ID__username=json_data["username"], absence_date_end=date-timedelta(days=1))[0]
-                    if ab_1.half_day == "NORMAL" and ab_2.half_day == "NORMAL":
-                        absence = Absence()
-                        absence.absence_date_start = ab_2.absence_date_start
-                        absence.absence_date_end = ab_1.absence_date_end
-                        absence.Target_User_ID_id = User.objects.get(username=json_data["username"]).id
-                        absence.User_ID = request.user
-                        ab_1.delete()
-                        ab_2.delete()
-                        absence.save()
-                    else:
-                        absence = non_connected()
-                    
-                elif date - timedelta(days=1) in Absence.objects.filter(Target_User_ID__username=json_data["username"]).values_list("absence_date_start", flat=True):
-                    a = Absence.objects.filter(Target_User_ID__username=json_data["username"], absence_date_start=date-timedelta(days=1))[0]
-                    if a.half_day == "NORMAL":
-                        a.absence_date_end = date
-                        a.save()
-                        absence = a
-                    else:
-                        absence = non_connected()
-                elif date + timedelta(days=1) in Absence.objects.filter(Target_User_ID__username=json_data["username"]).values_list("absence_date_end", flat=True):
-                    a = Absence.objects.filter(Target_User_ID__username=json_data["username"], absence_date_end=date+timedelta(days=1))[0]
-                    if a.half_day == "NORMAL":
-                        a.absence_date_start = date
-                        a.save()
-                        absence = a
-                    else:
-                        absence = non_connected()
-                elif date - timedelta(days=1) in Absence.objects.filter(Target_User_ID__username=json_data["username"]).values_list("absence_date_end", flat=True):
-                    a =Absence.objects.filter(Target_User_ID__username=json_data["username"], absence_date_end=date-timedelta(days=1))[0]
-                    if a.half_day == "NORMAL":
-                        a.absence_date_end = date
-                        a.save()
-                        absence = a
-                    else:
-                        absence = non_connected()
-                elif date + timedelta(days=1) in Absence.objects.filter(Target_User_ID__username=json_data["username"]).values_list("absence_date_start", flat=True):
-                    a = Absence.objects.filter(Target_User_ID__username=json_data["username"], absence_date_start=date+timedelta(days=1))[0]
-                    if a.half_day == "NORMAL":
-                        a.absence_date_start = date
-                        a.save()
-                        absence = a
-                    else:
-                        absence = non_connected()
-                else:
-                    absence = non_connected()
-
-                return JsonResponse({'start_date': absence.absence_date_start, 'end_date': absence.absence_date_end, 'taget_id': absence.Target_User_ID.username, 'user_id': absence.User_ID.username})
-        else:
-            return JsonResponse({})
-    else:
-        return HttpResponse('404')
-@login_required
-def click_remove(request):
-    if request.method == "POST":
-        data = json.loads(request.body)
-        date = datetime.datetime.strptime(data["date"], "%Y-%m-%d").date()
-        if UserProfile.objects.filter(user__username=data["username"]).exists():
-            perm_list = UserProfile.objects.filter(user__username=data["username"])[0].edit_whitelist.all()
-        else:
-            perm_list = [UserProfile.objects.get(user=request.user)]
-
-        if request.user in perm_list:
-            #Add an exception for recurring absence
-            if data["absence_type"] == "R":
-                exception = RecurringException()
-                exception.Target_User_ID = User.objects.get(username=data["username"])
-                exception.User_ID = request.user
-                exception.Exception_Start = data["date"]
-                exception.Exception_End = data["date"]
-                exception.save()
-            else:
-                #Remove absence if start date and end date is the same
-                if date in Absence.objects.filter(Target_User_ID__username=data["username"]).values_list("absence_date_start", flat=True) \
-                    and date in Absence.objects.filter(Target_User_ID__username=data["username"]).values_list("absence_date_end", flat=True):
-                    absence = Absence.objects.filter(Target_User_ID__username=data["username"], absence_date_start=date, absence_date_end=date)[0]
-                    absence.delete()
-                #Change absence start date if current start date removed
-                elif date in Absence.objects.filter(Target_User_ID__username=data["username"]).values_list("absence_date_start", flat=True):
-                    absence = Absence.objects.filter(Target_User_ID__username=data["username"], absence_date_start=date)[0]
-                    absence.absence_date_start = date + timedelta(days=1)
-                    absence.save()
-                #Change absence end date if current end date removed
-                elif date in Absence.objects.filter(Target_User_ID__username=data["username"]).values_list("absence_date_end", flat=True):
-                    absence = Absence.objects.filter(Target_User_ID__username=data["username"], absence_date_end=date)[0]
-                    absence.absence_date_end = date - timedelta(days=1)
-                    absence.save()
-                else:
-                    for absence in Absence.objects.filter(Target_User_ID__username=data["username"]):
-                        start_date = absence.absence_date_start
-                        end_date = absence.absence_date_end
-                        if date > start_date and date < end_date:
-                            ab_1 = Absence()
-                            ab_1.absence_date_start = start_date
-                            ab_1.absence_date_end = date - timedelta(days=1)
-                            ab_1.Target_User_ID_id = User.objects.get(username=data["username"]).id
-                            ab_1.User_ID = request.user
-
-                            ab_2 = Absence()
-                            ab_2.absence_date_start = date + timedelta(days=1)
-                            ab_2.absence_date_end = end_date
-                            ab_2.Target_User_ID_id = User.objects.get(username=data["username"]).id
-                            ab_2.User_ID = request.user
-
-                            absence.delete()
-                            ab_1.save()
-                            ab_2.save()
-
-        return JsonResponse({"start_date": data["date"]})
-    else:
-        return HttpResponse("404")
-
-@login_required
-def add_recurring(request) -> render:
-    if request.method == "POST":
-        form = RecurringAbsencesForm(request.POST)
-        form2 = TargetUserForm(request.POST, target_user=request.user)
-        form2.fields["target_user"].queryset = UserProfile.objects.filter(
-            edit_whitelist__in=[request.user]
-        )
-        rule = str(form["Recurrences"].value())
-
-
-        if not ("DAILY" in rule or "BY" in rule):
-            content = {
-                "form": form,
-                "form2": form2,
-                "message": "Must select a day/month",
-            }
-            return render(request, "ap_app/add_recurring_absence.html", content)
-
-        if form2.is_valid():
-            RecurringAbsences.objects.create(
-                Recurrences=form["Recurrences"].value(),
-                Target_User_ID=form2.cleaned_data["target_user"].user,
-                User_ID=request.user,
-            )
-            return redirect("index")
-    else:
-        form = RecurringAbsencesForm()
-        form2 = TargetUserForm(target_user=request.user)
-        form2.fields["target_user"].queryset = UserProfile.objects.filter(
-            edit_whitelist__in=[request.user]
-        )
-
-    content = {"form": form, "form2": form2}
-    return render(request, "ap_app/add_recurring_absence.html", content)
-
 
 @login_required
 def details_page(request) -> render:
@@ -670,458 +152,6 @@ def details_page(request) -> render:
     # TODO: get employee details and add them to context
     context = {"employee_dicts": ""}
     return render(request, "ap_app/Details.html", context)
-
-
-def get_date_data(
-    region,
-    month=datetime.datetime.now().strftime("%B"),
-    year=datetime.datetime.now().year,
-):
-    #  uses a dictionary to get all the data needed for the context
-    #  and concatenates it to form the full context with other dictionaries
-    data = {}
-    data["current_year"] = datetime.datetime.now().year
-    data["current_month"] = datetime.datetime.now().strftime("%B")
-    data["current_month_num"] = datetime.datetime.now().month
-    data["today"] = datetime.datetime.now().day
-    data["year"] = year
-    data["month"] = month
-    data["next_current_year"] = datetime.datetime.now().year + 1
-    data["next_second_year"] = datetime.datetime.now().year + 2
-    
-    start_date, end_date = "2022-07-03", "2024-07-03"
- 
-    data["month_list"] = pd.period_range(start=start_date, end=end_date, freq='M')
-    data["month_list"] = [month.strftime("%B %Y") for month in data["month_list"]]
-
-    data["selected_date"] = datetime.date(int(data["year"]), datetime.datetime.strptime(month, "%B").month, 1).strftime("%B %Y")
-
-    data["day_range"] = range(
-        1,
-        calendar.monthrange(
-            data["year"], datetime.datetime.strptime(month, "%B").month
-        )[1]
-        + 1,
-    )
-    data["day_range_num"] = len(list(data["day_range"])) + 1
-    data["month_num"] = datetime.datetime.strptime(month, "%B").month
-
-    data["previous_month"] = "December"
-    data["next_month"] = "January"
-    data["previous_year"] = year - 1
-    data["next_year"] = year + 1
-
-    # as the month number resets every year try and except statements
-    # have to be used as at the end and start of a year
-    # the month cannot be calculated by adding or subtracting 1
-    # as 13 and 0 are not datetime month numbers
-    try:
-        data["next_month"] = datetime.datetime.strptime(
-            str((datetime.datetime.strptime(data["month"], "%B")).month + 1), "%m"
-        ).strftime("%B")
-
-    except ValueError:
-        pass
-    try:
-        data["previous_month"] = datetime.datetime.strptime(
-            str((datetime.datetime.strptime(data["month"], "%B")).month - 1), "%m"
-        ).strftime("%B")
-    except ValueError:
-        pass
-
-    # calculating which days are weekends to mark them easier in the html
-    data["days_name"] = []
-    for day in data["day_range"]:
-        date = f"{day} {month} {year}"
-        date = datetime.datetime.strptime(date, "%d %B %Y")
-        date = date.strftime("%A")[0:2]
-        data["days_name"].append(date)
-
-    data["weekend_list"] = []
-    for day in data["day_range"]:
-        date = f"{day} {month} {year}"
-        date = datetime.datetime.strptime(date, "%d %B %Y")
-        date = date.strftime("%A")[0:2]
-        if (date == "Sa" or date == "Su"):
-            data["weekend_list"].append(day)
-
-    data["bank_hol"] = []
-    for h in holidays.country_holidays(region, years=year).items():
-        if (h[0].month == data["month_num"]):
-            data["bank_hol"].append(h[0].day)
-        
-    return data
-
-
-
-
-def get_absence_data(users, user_type):
-    data = {}
-    absence_content = []
-    total_absence_dates = {}
-    total_half_dates = {}
-    total_recurring_dates = {}
-    all_absences = {}
-    delta = datetime.timedelta(days=1)
-
-    for user in users:
-        # all the absences for the user
-        if user_type == 1:
-            user_username = user.user.username
-        else:
-            user_username = user.username
-
-        absence_info = Absence.objects.filter(Target_User_ID__username=user_username)
-        total_absence_dates[user_username] = []
-        total_recurring_dates[user_username] = []
-        total_half_dates[user_username] = []
-        all_absences[user_username] = []
-
-        # if they have any absences
-        if absence_info:
-            # mapping the absence content to keys in dictionary
-            for x in absence_info:
-                absence_id = x.ID
-                absence_date_start = x.absence_date_start
-                absence_date_end = x.absence_date_end
-                dates = absence_date_start
-                if x.half_day == "NORMAL":
-                    while dates <= absence_date_end:
-                        total_absence_dates[user_username].append(dates)
-                        dates += delta
-                else:
-                    total_half_dates[user_username].append(
-                        {
-                            "date": absence_date_start,
-                            "type": x.half_day
-                        }
-                    )
-
-                absence_content.append(
-                    {
-                        "ID": absence_id,
-                        "absence_date_start": absence_date_start,
-                        "absence_date_end": absence_date_end,
-                        "dates": total_absence_dates[user_username],
-                    }
-                )
-
-            # for each user it maps the set of dates to a dictionary key labelled as the users name
-            all_absences[user_username] = absence_content
-
-        recurring = RecurringAbsences.objects.filter(Target_User_ID__username=user_username)
-
-        if recurring:
-            for recurrence_ in recurring:
-                dates = recurrence_.Recurrences.occurrences(
-                    dtend=datetime.datetime.strptime(
-                        str(datetime.datetime.now().year + 2), "%Y"
-                    )
-                )
-
-                for x in list(dates)[:-1]:
-                    time_const = "23:00:00"
-                    time_var = datetime.datetime.strftime(x, "%H:%M:%S")
-                    if time_const == time_var:
-                        x += timedelta(days=1)
-                    
-                    #print(RecurringException.objects.filter(Target_User_ID__username=user_username, Exception_Start=x).count())
-                    if RecurringException.objects.filter(Target_User_ID__username=user_username, Exception_Start=x).count() == 0:
-                        total_recurring_dates[user_username].append(x)
-                # TODO: add auto deleting for recurring absences once last date of absences in before now
-                # if x < datetime.datetime.now():
-                #    pass
-
-    data["recurring_absence_dates"] = total_recurring_dates
-    data["all_absences"] = all_absences
-    data["absence_dates"] = total_absence_dates
-    data["half_days_data"] = total_half_dates
-    data["users"] = users
-    return data
-
-
-@login_required
-def team_calendar(
-    request,
-    id,
-    month=datetime.datetime.now().strftime("%B"),
-    year=datetime.datetime.now().year,
-):
-    if is_member(request.user, id):
-
-        # Get acceptable date - (NOTHING BELOW now - 12months)
-        date = check_calendar_date(year, month)
-        if date:
-            month = date.strftime("%B")
-            year = date.year
-
-        try:
-            userprofile: UserProfile = UserProfile.objects.filter(user=request.user)[0]
-        except IndexError:
-            return redirect("/")
-
-        data_1 = get_date_data(userprofile.region, month, year)
-
-        users = Relationship.objects.all().filter(
-            team=id, status=Status.objects.get(status="Active")
-        )
-
-        data_2 = get_absence_data(users, 1)
-
-        team = Team.objects.get(id=id)
-
-        # Filtering users by privacy
-        filtered_users = []
-        viewer = data_2["users"].get(user=request.user)
-        if str(viewer.role) == "Follower":
-            # Than hide data of users with privacy on
-            for user in data_2["users"]:
-                user_profile = UserProfile.objects.get(user=user.user)
-                if not user_profile.privacy:
-                    filtered_users.append(user)
-
-                else:
-                    # For pop-up to inform viewer that there are hidden users
-                    data_2.update({"hiding_users": True})
-
-            data_2["users"] = filtered_users
-
-        # Gets filtered users by filtering system on page
-        filtered_users = get_filter_users(
-            request, [user.user for user in data_2["users"]]
-        )
-
-        actual_filtered_users = []
-        for user in data_2["users"]:
-            if user.user in filtered_users:
-                actual_filtered_users.append(user)
-
-        # Reconstructs users list to be in desired form for template - (QuerySet of relationships)
-        data_2["users"] = actual_filtered_users
-
-        user_in_teams = []
-        for rel in Relationship.objects.filter(team=team):
-            user_in_teams.append(rel.user.id)
-
-        data_3 = {
-            "owner": Role.objects.all()[0],
-            "Sa": "Sa",
-            "Su": "Su",
-            "current_user": Relationship.objects.get(user=request.user, team=team),
-            "team": team,
-            "all_users": User.objects.all().exclude(id__in=user_in_teams),
-            "team_count": Relationship.objects.filter(
-                team=team.id, status=Status.objects.get(status="Active")
-            ).count(),
-        }
-
-        context = {**data_1, **data_2, **data_3}
-
-        return render(request, "teams/calendar.html", context)
-
-    return redirect("dashboard")
-
-
-@login_required
-def all_calendar(
-    request,
-    month=datetime.datetime.now().strftime("%B"),
-    year=datetime.datetime.now().year,
-):
-    # Get acceptable date - (NOTHING BELOW now - 12months)
-    date = check_calendar_date(year, month)
-    if date:
-        month = date.strftime("%B")
-        year = date.year
-    
-    try:
-        userprofile: UserProfile = UserProfile.objects.filter(user=request.user)[0]
-    except IndexError:
-        return redirect("/")
-    
-    if userprofile.external_teams:
-        try:
-            token = (str(request.user) + "AbsencePlanner").encode()
-            encryption = hashlib.sha256(token).hexdigest()
-            if requests.get(env("TEAM_DATA_URL") + "api/teams/?username={}".format(request.user.username), headers={"TEAMS-TOKEN": encryption}).status_code == 200:
-                return redirect("/calendar/1")
-        except:
-            print("Failed to load api")
-
-
-    data_1 = get_date_data(userprofile.region, month, year)
-    
-    current_month = data_1["current_month"]
-    current_year = data_1["current_year"]
-    current_day = datetime.datetime.now().day
-    date = f"{current_day} {current_month} {current_year}"
-    date = datetime.datetime.strptime(date, "%d %B %Y")
-
-
-    all_users = []
-    all_users.append(request.user)
-
-    user_relations = Relationship.objects.filter(
-        user=request.user, status=Status.objects.get(status="Active")
-    )
-    hiding_users = False
-
-
-    for index, relation in enumerate(user_relations):
-        rels = Relationship.objects.filter(
-            team=relation.team, status=Status.objects.get(status="Active")
-        )
-
-        # Finds the viewers role in the team
-        for user in rels:
-            if user.user == request.user:
-                viewers_role = Role.objects.get(id=user.role_id)
-
-        for rel in rels:
-            if rel.user not in all_users:
-                if viewers_role.role == "Follower":
-                    # Than hide users data who have privacy on
-
-                    user_profile = UserProfile.objects.get(user=rel.user)
-                    if not user_profile.privacy:
-                        # If user hasn't got their data privacy on
-                        all_users.append(rel.user)
-                    else:
-                        # Used to inform user on calendar page if there are hiden users
-                        hiding_users = True
-
-                else:
-                    # Only followers cannot view those who have privacy set for their data. - (Members & Owners can see the data)
-                    all_users.append(rel.user)
-
-    # Filtering
-    filtered_users = get_filter_users(request, all_users)
-
-    data_2 = get_absence_data(all_users, 2)
-
-    data_3 = {"Sa": "Sa", "Su": "Su", "users_filter": filtered_users}
-
-  
-    grid_calendar_month_values = list(data_1["day_range"])
-    # NOTE: This will select which value to use to fill in how many blank cells where previous month overrides prevailing months days. 
-    # This is done by finding the weekday value for the 1st day of the month. Example: "Tu" will require 1 blank space/cell.
-    for i in range({"Mo":0, "Tu":1, "We":2, "Th":3, "Fr":4, "Sa":5, "Su":6}[data_1["days_name"][0]]):
-        grid_calendar_month_values.insert(0, -1)
-
-
-    context = {
-        **data_1,
-        **data_2,
-        **data_3,
-        "users_hidden": hiding_users,
-        "home_active": True,
-        
-        # Grid-Calendars day values  
-        "detailed_calendar_day_range":grid_calendar_month_values,
-     
-        
-    }
-   
-
-    return render(request, "ap_app/calendar.html", context)
-
-@login_required
-def set_calendar_month(request):
-    if request.method == "POST":
-        month = request.POST.get('month_names')
-        
-        split = month.split()
-
-        month = split[0]
-
-        year = split[1]
-
-        print(month)
-
-    return redirect('all_calendar', month=month, year=year)
-
-
-def text_rules(user):
-    recurring_absences = RecurringAbsences.objects.filter(Target_User_ID=user).values(
-        "Recurrences", "ID"
-    )
-    rec_absences = {}
-
-    for x in recurring_absences:
-        absence_ = x["Recurrences"]
-        if absence_:
-            rec_absences[x["ID"]] = []
-        if absence_.exdates:
-            for y in absence_.exdates:
-                rec_absences[x["ID"]].append(
-                    "Excluding Date: " + (y + timedelta(days=1)).strftime("%A,%d %B,%Y")
-                )
-        if absence_.rdates:
-            for y in absence_.rdates:
-                rec_absences[x["ID"]].append(
-                    "Date: " + (y + timedelta(days=1)).strftime("%A,%d %B,%Y")
-                )
-        if absence_.rrules:
-            for y in absence_.rrules:
-                rec_absences[x["ID"]].append("Rule: " + str(y.to_text()))
-
-        if absence_.exrules:
-            for y in absence_.exrules:
-                rec_absences[x["ID"]].append("Excluding Rule: " + str(y.to_text()))
-
-    return rec_absences
-
-
-# Profile page
-@login_required
-def profile_page(request):
-    if request.method == "POST":
-        form = SwitchUser(
-            request.POST,
-            initial={"user": request.user},
-        )
-        form.fields["user"].queryset = UserProfile.objects.filter(
-            edit_whitelist=request.user
-        )
-        users = UserProfile.objects.filter(edit_whitelist=request.user)
-        rec_absences = text_rules(request.user)
-
-        if form.is_valid():
-            absence_user = form.cleaned_data["user"].user
-
-            absences = Absence.objects.filter(Target_User_ID=absence_user)
-            rec_absences = text_rules(absence_user)
-            return render(
-                request,
-                "ap_app/profile.html",
-                {
-                    "form": form,
-                    "message": "Successfully switched user",
-                    "absences": absences,
-                    "users": users,
-                    "recurring_absences": rec_absences,
-                },
-            )
-    else:
-        absences = Absence.objects.filter(Target_User_ID=request.user.id)
-        rec_absences = text_rules(request.user)
-
-        users = UserProfile.objects.filter(edit_whitelist=request.user)
-        form = SwitchUser()
-        form.fields["user"].queryset = users
-        form.fields["user"].initial = request.user
-
-        return render(
-            request,
-            "ap_app/profile.html",
-            {
-                "absences": absences,
-                "users": users,
-                "form": form,
-                "recurring_absences": rec_absences,
-            },
-        )
-
 
 @login_required
 def deleteuser(request):
@@ -1138,80 +168,6 @@ def deleteuser(request):
     context = {"delete_form": delete_form}
 
     return render(request, "registration/delete_account.html", context)
-
-
-@login_required
-def absence_delete(request, absence_id: int, user_id: int, team_id: int = 1):
-    try:
-        absence = Absence.objects.get(pk=absence_id)
-        absence.delete()
-    except Absence.DoesNotExist: 
-        pass
-    if request.user == User.objects.get(id = user_id):
-        return redirect("profile")
-    return redirect("edit_team_member_absence", team_id, user_id)
-
-
-@login_required
-def absence_recurring_delete(
-    request, absence_id: int, user_id: int, team_id: int = None
-):
-    absence = RecurringAbsences.objects.get(pk=absence_id)
-    user = request.user
-    absence.delete()
-    if user == absence.Target_User_ID:
-        return redirect("profile")
-    return redirect("edit_team_member_absence", team_id, user_id)
-
-
-class EditAbsence(UpdateView):
-    template_name = "ap_app/edit_absence.html"
-    model = Absence
-
-    # specify the fields
-    fields = ["absence_date_start", "absence_date_end"]
-
-    def get_success_url(self) -> str:
-        return reverse("profile")
-
-
-@login_required
-def edit_recurring_absences(request, pk):
-    absence = RecurringAbsences.objects.get(ID=pk)
-
-    if request.method == "POST":
-        form = RecurringAbsencesForm(request.POST, instance=absence)
-        form2 = TargetUserForm(request.POST, target_user=request.user)
-        form2.fields["target_user"].queryset = UserProfile.objects.filter(
-            edit_whitelist__in=[request.user]
-        )
-        rule = str(form["Recurrences"].value())
-
-        if not ("DAILY" in rule or "BY" in rule):
-            content = {
-                "form": form,
-                "form2": form2,
-                "message": "Must select a day/month",
-            }
-            return render(request, "ap_app/edit_recurring_absence.html", content)
-
-        if form2.is_valid():
-            absence.Target_User_ID = form2.cleaned_data["target_user"].user
-            absence.recurrences = form["Recurrences"].value()
-            absence.save()
-            return redirect("index")
-    else:
-        form = RecurringAbsencesForm(instance=absence)
-        
-        form2 = TargetUserForm(target_user=absence.Target_User_ID)
-        form2.fields["target_user"].queryset = UserProfile.objects.filter(
-            edit_whitelist__in=[absence.User_ID]
-        )
-
-    return render(
-        request, "ap_app/edit_recurring_absence.html", {"form": form, "form2": form2}
-    )
-
 
 @login_required
 def profile_settings(request) -> render:
@@ -1243,18 +199,13 @@ def profile_settings(request) -> render:
             userprofile.privacy = False
         elif request.POST.get("privacy") == "on":
             userprofile.privacy = True
-        if request.POST.get("teams") == None:
-            userprofile.external_teams = False
-        elif request.POST.get("teams") == "on":
-            userprofile.external_teams = True
         userprofile.save()
     
     country_data = get_region_data()
     country_name = pycountry.countries.get(alpha_2=userprofile.region).name
 
     privacy_status = userprofile.privacy
-    teams_status = userprofile.external_teams
-    context = {"userprofile": userprofile, "data_privacy_mode": privacy_status, "external_teams": teams_status,"current_country": country_name, **country_data}
+    context = {"userprofile": userprofile, "data_privacy_mode": privacy_status, "current_country": country_name, **country_data}
     return render(request, "ap_app/settings.html", context)
 
 
@@ -1313,220 +264,138 @@ def set_region(request):
 
     return redirect("/profile/settings")
 
-def find_user_obj(user_to_find):
-    """Finds & Returns object of 'UserProfile' for a user
-    \n-param (type)User user_to_find
-    """
-    users = UserProfile.objects.filter(user=user_to_find)
-    # If cannot find object for a user, than creates on
-
-    if users.count() <= 0:
-        UserProfile.objects.create(user=user_to_find, accepted_policy=False)
-        user_found = UserProfile.objects.filter(user=user_to_find)[0]
-        user_found.edit_whitelist.add(user_to_find)
-
-    # Users object
-    user_found = UserProfile.objects.filter(user=user_to_find)[0]
-
-    return user_found
-
-
-def obj_exists(user):
-    """Determines if a user has a 'UserProfile' Object"""
-    objs = UserProfile.objects.filter(user=user)
-    if objs.count() == 0:
-        return False
-
-    return True
-
-
-def get_filter_users(request, users) -> list:
-    """Used for calendar filtering system - (Returns list of filtered users depending on
-    search-bar and 'filter by absence' checkbox"""
-
-    filtered_users = []
-
-    # Filtering by both username & absence
-    if "username" in request.GET and "absent" in request.GET:
-        # Get username input & limits the length to 50
-        name_filtered_by = request.GET["username"][:50]
-        for absence in Absence.objects.all():
-            user = User.objects.get(id=absence.Target_User_ID.id)
-            if user in users:
-                username = user.username
-                if (
-                    absence.Target_User_ID not in filtered_users
-                    and name_filtered_by.lower() in username.lower()
-                ):
-                    filtered_users.append(absence.Target_User_ID)
-
-    # ONLY filtering by username
-    elif "username" in request.GET:
-        # Name limit is 50
-        name_filtered_by = request.GET["username"][:50]
-        for user in users:
-            # same logic as "icontains", searches through users names & finds similarities
-            if name_filtered_by.lower() in user.username.lower():
-                filtered_users.append(user)
-
-    # ONLY filtering by absences
-    elif "absent" in request.GET:
-        for absence in Absence.objects.all():
-            user = User.objects.get(id=absence.Target_User_ID.id)
-            if user in users and absence.Target_User_ID not in filtered_users:
-                filtered_users.append(absence.Target_User_ID)
-
-    # Else, no filtering
-    else:
-        filtered_users = users
-
-    return filtered_users
-
-
-def is_member(user, team_id) -> bool:
-    """ Determines if the user is a member of the team before accessing its contents """
-
-    team = Relationship.objects.filter(team=Team.objects.get(id=team_id))
-
-    # Boolean determines if viewer is in this team
-    for rel in team:
-        if rel.user == user:
-            return True
-
-    # Else user has changed URL & is attempting to view other teams content
-    return False
-
-
-def is_owner(user, team_id) -> bool:
-    """ Determines if the user is an owner of the team """
-    
-    user_relation = Relationship.objects.get(team=team_id, user=user)
-    return (user_relation.role.role == "Owner")
-
-
-def check_calendar_date(year, month) -> datetime.datetime:
-    """ This function will determine if the requested date is acceptable - (NOT before current date - 12 months) """
-    # Current year * 12     + current month as num      = amount of months been  
-    # requested year * 12   + requested month as num    = 
-    
-    # If requested date is before "current date - 12 months" than will not accept date
-    todays_date = datetime.datetime.now()
-    requested_months_amount = (int(year) * 12) + datetime.datetime.strptime(month, "%B").month
-    current_months_amount   = (todays_date.year * 12) + todays_date.month 
-    months_difference = current_months_amount - requested_months_amount
-
-
-    if months_difference > 12:
-        # Return the most earliest date acceptable - (now - 12 months)
-        return datetime.datetime(todays_date.year - 1, todays_date.month, 1)  
-    else:
-        return None
-
-from django.shortcuts import render
-
-def my_view(request):
-    return render(request, "base.html")
-
-
-def handler404(request, exception):
-    context = {}
-    response = render(request, "404.html", context=context)
-    response.status_code = 404
-    return response
-
-def handler500(request):
-    context = {}
-    response = render(request, "500.html", context=context)
-    response.status_code = 500
-    return response
-
-def handler400(request, exception):
-    context = {}
-    response = render(request, "400.html", context=context)
-    response.status_code = 400
-    return response
-
-
 #JC - Calendar view using the API
+
+#Add an absence when clicking on the calendar
 @login_required
-def api_calendar_view(
-    request,
-    #JC - These are the default values for the calendar.
-    month=datetime.datetime.now().strftime("%B"),
-    year=datetime.datetime.now().year
-):
-    
-    try:
-        userprofile: UserProfile = UserProfile.objects.get(user=request.user)
-    except IndexError:
-        return redirect("/")
-    
-    if not userprofile.external_teams:
-        return redirect("/calendar/0")
-    
-    #JC - Get API data
-    api_data = None
-    if request.method == "GET":
-        try:
-            token = (str(request.user) + "AbsencePlanner").encode()
-            encryption = hashlib.sha256(token).hexdigest()
-            r = requests.get(env("TEAM_DATA_URL") + "api/teams/?username={}".format(request.user.username), headers={"TEAMS-TOKEN": encryption})
-        except:
-            print("API failed to connect")
-            return redirect("/")
-        if r.status_code == 200:
-            api_data = r.json()
-        else:
-            if r:
-                result = r.json()
-                if result["code"] == "I":
-                    print("Username not found in Team App database.")
-                elif result["code"] == "N":
-                    print("A username was not provided with the request.")
+def click_add(request):
+    if request.method == "POST":
+        json_data=json.loads(request.body)
+        perm_list = UserProfile.objects.filter(id=json_data["id"])[0].edit_whitelist.all()
+        if request.user in perm_list:
+            date = datetime.datetime.strptime(json_data["date"], "%Y-%m-%d").date()
+            #This will add a half
+            if json_data["half_day"] == True:
+                absence = Absence()
+                absence.absence_date_start = json_data['date']
+                absence.absence_date_end = json_data['date']
+                absence.Target_User_ID_id = json_data["id"]
+                absence.User_ID = request.user
+                if json_data["half_day_time"] == "M":
+                    absence.half_day = "MORNING"
+                elif json_data["half_day_time"] == "A":
+                    absence.half_day = "AFTERNOON"
+                absence.save()
+                return JsonResponse({})
             else:
-                print("Fatal Error")
+                def non_connected():
+                    absence = Absence()
+                    absence.absence_date_start = json_data['date']
+                    absence.absence_date_end = json_data['date']
+                    absence.Target_User_ID_id = json_data["id"]
+                    absence.User_ID = request.user
+                    absence.save()
+                    return absence
+                date = datetime.datetime.strptime(json_data["date"], "%Y-%m-%d").date()
+                absence = None
+                if date - timedelta(days=1) in Absence.objects.filter(Target_User_ID_id=json_data["id"]).values_list("absence_date_end", flat=True) \
+                    and date + timedelta(days=1) in Absence.objects.filter(Target_User_ID_id=json_data["id"]).values_list("absence_date_start", flat=True):
+                    ab_1 = Absence.objects.filter(Target_User_ID_id=json_data["id"], absence_date_start=date+timedelta(days=1))[0]
+                    ab_2 = Absence.objects.filter(Target_User_ID_id=json_data["id"], absence_date_end=date-timedelta(days=1))[0]
+                    if ab_1.half_day == "NORMAL" and ab_2.half_day == "NORMAL":
+                        absence = Absence()
+                        absence.absence_date_start = ab_2.absence_date_start
+                        absence.absence_date_end = ab_1.absence_date_end
+                        absence.Target_User_ID_id = json_data["id"]
+                        absence.User_ID = request.user
+                        ab_1.delete()
+                        ab_2.delete()
+                        absence.save()
+                    else:
+                        absence = non_connected()
+                    
+                elif date - timedelta(days=1) in Absence.objects.filter(Target_User_ID_id=json_data["id"]).values_list("absence_date_start", flat=True):
+                    a = Absence.objects.filter(Target_User_ID_id=json_data["id"], absence_date_start=date-timedelta(days=1))[0]
+                    if a.half_day == "NORMAL":
+                        a.absence_date_end = date
+                        a.save()
+                        absence = a
+                    else:
+                        absence = non_connected()
+                elif date + timedelta(days=1) in Absence.objects.filter(Target_User_ID_id=json_data["id"]).values_list("absence_date_end", flat=True):
+                    a = Absence.objects.filter(Target_User_ID_id=json_data["id"], absence_date_end=date+timedelta(days=1))[0]
+                    if a.half_day == "NORMAL":
+                        a.absence_date_start = date
+                        a.save()
+                        absence = a
+                    else:
+                        absence = non_connected()
+                elif date - timedelta(days=1) in Absence.objects.filter(Target_User_ID_id=json_data["id"]).values_list("absence_date_end", flat=True):
+                    a =Absence.objects.filter(Target_User_ID_id=json_data["id"], absence_date_end=date-timedelta(days=1))[0]
+                    if a.half_day == "NORMAL":
+                        a.absence_date_end = date
+                        a.save()
+                        absence = a
+                    else:
+                        absence = non_connected()
+                elif date + timedelta(days=1) in Absence.objects.filter(Target_User_ID_id=json_data["id"]).values_list("absence_date_start", flat=True):
+                    a = Absence.objects.filter(Target_User_ID_id=json_data["id"], absence_date_start=date+timedelta(days=1))[0]
+                    if a.half_day == "NORMAL":
+                        a.absence_date_start = date
+                        a.save()
+                        absence = a
+                    else:
+                        absence = non_connected()
+                else:
+                    absence = non_connected()
 
-    date = check_calendar_date(year, month)
-    if date:
-        month = date.strftime("%B")
-        year = date.year
+                return JsonResponse({'start_date': absence.absence_date_start, 'end_date': absence.absence_date_end, 'taget_id': absence.Target_User_ID.username, 'user_id': absence.User_ID.username})
+        else:
+            return JsonResponse({})
+    else:
+        return HttpResponse('404')
+@login_required
+def click_remove(request):
+    if request.method == "POST":
+        data = json.loads(request.body)
+        date = datetime.datetime.strptime(data["date"], "%Y-%m-%d").date()
+        #Remove absence if start date and end date is the same
+        if date in Absence.objects.filter(Target_User_ID_id=data["id"]).values_list("absence_date_start", flat=True) \
+            and date in Absence.objects.filter(Target_User_ID_id=data["id"]).values_list("absence_date_end", flat=True):
+            absence = Absence.objects.filter(Target_User_ID_id=data["id"], absence_date_start=date, absence_date_end=date)[0]
+            absence.delete()
+        #Change absence start date if current start date removed
+        elif date in Absence.objects.filter(Target_User_ID_id=data["id"]).values_list("absence_date_start", flat=True):
+            absence = Absence.objects.filter(Target_User_ID_id=data["id"], absence_date_start=date)[0]
+            absence.absence_date_start = date + timedelta(days=1)
+            absence.save()
+        #Change absence end date if current end date removed
+        elif date in Absence.objects.filter(Target_User_ID_id=data["id"]).values_list("absence_date_end", flat=True):
+            absence = Absence.objects.filter(Target_User_ID_id=data["id"], absence_date_end=date)[0]
+            absence.absence_date_end = date - timedelta(days=1)
+            absence.save()
+        else:
+            for absence in Absence.objects.filter(Target_User_ID_id=data["id"]):
+                start_date = absence.absence_date_start
+                end_date = absence.absence_date_end
+                if date > start_date and date < end_date:
+                    ab_1 = Absence()
+                    ab_1.absence_date_start = start_date
+                    ab_1.absence_date_end = date - timedelta(days=1)
+                    ab_1.Target_User_ID_id = data["id"]
+                    ab_1.User_ID = request.user
 
-    data_1 = get_date_data(userprofile.region, month, year)
+                    ab_2 = Absence()
+                    ab_2.absence_date_start = date + timedelta(days=1)
+                    ab_2.absence_date_end = end_date
+                    ab_2.Target_User_ID_id = data["id"]
+                    ab_2.User_ID = request.user
+
+                    absence.delete()
+                    ab_1.save()
+                    ab_2.save()
+
+        return JsonResponse({"start_date": data["date"]})
+    else:
+        return HttpResponse("404")
     
-    current_month = data_1["current_month"]
-    current_year = data_1["current_year"]
-    current_day = datetime.datetime.now().day
-    date = f"{current_day} {current_month} {current_year}"
-    date = datetime.datetime.strptime(date, "%d %B %Y")
-
-
-    all_users = []
-    all_users.append(request.user)
-
-    if api_data:
-        for team in api_data:
-            for member in team["team"]["members"]:
-                retrieved_user = User.objects.filter(username=member["user"]["username"])
-                if retrieved_user.exists() and retrieved_user not in all_users:
-                    all_users.append(retrieved_user[0])
-
-    data_2 = get_absence_data(all_users, 2)
-
-    data_3 = {"Sa": "Sa", "Su": "Su"}
-
-    grid_calendar_month_values = list(data_1["day_range"])
-    # NOTE: This will select which value to use to fill in how many blank cells where previous month overrides prevailing months days. 
-    # This is done by finding the weekday value for the 1st day of the month. Example: "Tu" will require 1 blank space/cell.
-    for i in range({"Mo":0, "Tu":1, "We":2, "Th":3, "Fr":4, "Sa":5, "Su":6}[data_1["days_name"][0]]):
-        grid_calendar_month_values.insert(0, -1)
-
-
-    context = {
-        **data_1,
-        **data_2,
-        **data_3,
-        "home_active": True,
-        "api_data": api_data,
-    }
-
-    return render(request, "api_pages/calendar.html", context)
