+++ ap_src/ap_site/example_env.txt
@@ -1,0 +1,1 @@
+TEAM_DATA_URL=http://localhost:8000/
